VCV-162: move getHierarchyLevels into site utils
