+++ src/features/operations/components/site-list/operations-site-hierarchy.tsx
@@ -24,28 +24,6 @@ function getCompletenessBackgroundColor(
     return 'bg-destructive/10 hover:bg-destructive/20';
 }
 
-const LEGACY_HIERARCHY_LEVELS = [
-    { key: 'subCounty', label: 'Subcounty' },
-    { key: 'healthCenter', label: 'Health Center' },
-    { key: 'parish', label: 'Parish' },
-    { key: 'villageName', label: 'Village' },
-    { key: 'houseNumber', label: 'House' },
-];
-
-export function getHierarchyLevels(
-    sites: Site[],
-): { key: string; label: string }[] {
-    const firstSite = sites[0];
-    if (!firstSite) return LEGACY_HIERARCHY_LEVELS;
-
-    const hierarchyKeys = Object.keys(firstSite.locationHierarchy);
-    if (hierarchyKeys.length > 0) {
-        return hierarchyKeys.slice(1).map(key => ({ key, label: key }));
-    }
-
-    return LEGACY_HIERARCHY_LEVELS;
-}
-
 function getSiteValueAtLevel(site: Site, levelKey: string): string | undefined {
     const hierarchyKeys = Object.keys(site.locationHierarchy);
     if (hierarchyKeys.length > 0) {
